pyproject.toml
[project]
name = "ntt-rag-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "docling>=2.48.0",
    "fastapi>=0.116.1",
    "huggingface-hub[hf-xet]>=0.34.4",
    "langchain>=0.3.27",
    "langchain-core>=0.3.75",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.32",
    "pydantic-ai>=0.7.6",
    "pytest>=8.4.1",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.28.1",
    "streamlit>=1.49.0",
    "anyio>=4.10.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
]
markers = [
    "asyncio: marks tests as async",
    "slow: marks tests as slow running",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "ingestion", "ui"]
